+++ src/core/injections/blind/techniques/time_based/tb_injector.py
@@ -71,7 +71,7 @@ def examine_requests(payload, vuln_parameter, http_request_method, url, timesec,
       payload = payload.replace("\"", "\\\"")
       data = parameter.replace(settings.INJECT_TAG, urllib.unquote(payload))
       try:
-        data = json.loads(data, strict = False)
+        data = json.loads(checks.json_data(data), strict = False)
       except:
         pass
       request = urllib2.Request(url, json.dumps(data))
@@ -127,7 +127,7 @@ def injection_test(payload, http_request_method, url):
       payload = payload.replace("\"", "\\\"")
       data = parameter.replace(settings.INJECT_TAG, urllib.unquote(payload))
       try:
-        data = json.loads(data, strict = False)
+        data = json.loads(checks.json_data(data), strict = False)
       except:
         pass
       request = urllib2.Request(url, json.dumps(data))