+++ src/core/injections/results_based/techniques/eval_based/eb_injector.py
@@ -75,7 +75,7 @@ def injection_test(payload, http_request_method, url):
       payload = payload.replace("\"", "\\\"")
       data = parameter.replace(settings.INJECT_TAG, urllib.unquote(payload))
       try:
-        data = json.loads(data, strict = False)
+        data = json.loads(checks.json_data(data), strict = False)
       except:
         pass
       request = urllib2.Request(url, json.dumps(data))
@@ -266,7 +266,7 @@ def injection(separator, TAG, cmd, prefix, suffix, whitespace, http_request_meth
           payload = payload.replace("\"", "\\\"")
           data = parameter.replace(settings.INJECT_TAG, urllib.unquote(payload))
           try:
-            data = json.loads(data, strict = False)
+            data = json.loads(checks.json_data(data), strict = False)
           except:
             pass
           request = urllib2.Request(url, json.dumps(data))