+++ src/core/injections/semiblind/techniques/tempfile_based/tfb_injector.py
@@ -73,7 +73,7 @@ def examine_requests(payload, vuln_parameter, http_request_method, url, timesec,
       payload = payload.replace("\"", "\\\"")
       data = parameter.replace(settings.INJECT_TAG, urllib.unquote(payload))
       try:
-        data = json.loads(data, strict = False)
+        data = json.loads(checks.json_data(data), strict = False)
       except:
         pass
       request = urllib2.Request(url, json.dumps(data))
@@ -130,7 +130,7 @@ def injection_test(payload, http_request_method, url):
       payload = payload.replace("\"", "\\\"")
       data = parameter.replace(settings.INJECT_TAG, urllib.unquote(payload))
       try:
-        data = json.loads(data, strict = False)
+        data = json.loads(checks.json_data(data), strict = False)
       except:
         pass
       request = urllib2.Request(url, json.dumps(data))