+++ src/core/injections/blind/techniques/time_based/tb_injector.py
@@ -70,10 +70,10 @@ def examine_requests(payload, vuln_parameter, http_request_method, url, timesec,
     if settings.IS_JSON:
       data = parameter.replace(settings.INJECT_TAG, urllib.unquote(payload.replace("\"", "\\\"")))
       try:
-        data = json.loads(checks.json_data(data), strict = False)
+        data = checks.json_data(data)
       except:
         pass
-      request = urllib2.Request(url, json.dumps(data))
+      request = urllib2.Request(url, data)
     else:
       if settings.IS_XML:
         data = parameter.replace(settings.INJECT_TAG, urllib.unquote(payload)) 
@@ -125,10 +125,10 @@ def injection_test(payload, http_request_method, url):
     if settings.IS_JSON:
       data = parameter.replace(settings.INJECT_TAG, urllib.unquote(payload.replace("\"", "\\\"")))
       try:
-        data = json.loads(checks.json_data(data), strict = False)
+        data = checks.json_data(data)
       except:
         pass
-      request = urllib2.Request(url, json.dumps(data))
+      request = urllib2.Request(url, data)
     else:
       if settings.IS_XML:
         data = parameter.replace(settings.INJECT_TAG, urllib.unquote(payload))