+++ src/core/injections/results_based/techniques/classic/cb_injector.py
@@ -99,7 +99,6 @@ def injection_test_results(response, TAG, randvcalc):
   if response == False:
     return False
   else:
-
     # Check the execution results
     html_data = response.read()
     html_data = html_data.replace("\n"," ")