+++ readme/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Version 2.9 (upcoming)
+* Added: GPL Cooperation Commitment (COMMITMENT.txt) has been added.
 * Updated: Minor update regarding HTTP authentication (Basic / Digest).
 * Revised: Minor improvements regarding preventing false negative results, due to parameters tampering during the detection phase.
 * Revised: Minor improvements regarding "reverse_tcp" and "bind_tcp" shell options.